+++ packages/apollos-data-prayer/src/prayer-requests/schema.js
@@ -24,7 +24,7 @@ const prayerRequestSchema = gql`
     firstName: String
     lastName: String
     text: String!
-    enteredDateTime: String
+    enteredDateTime: String!
     isPublic: Boolean
     flagCount: Int
     prayerCount: Int
