+++ packages/newspringchurchapp/src/tabs/connect/Connect.js
@@ -15,7 +15,7 @@ import {
   Icon,
   PaddedView,
 } from '@apollosproject/ui-kit';
-import MyPrayers from '../../ui/MyPrayer';
+import UserPrayers from '../../ui/UserPrayer';
 import ActionTable from './ActionTable';
 import Toolbar from './Toolbar';
 import { UserAvatarHeaderConnected } from './UserAvatarHeader';
@@ -74,7 +74,7 @@ class Connect extends PureComponent {
                     <RecentlyLikedTileFeedConnected key="RecentlyLikedTileFeedConnected" />
                     <Toolbar />
                     <ActionTable />
-                    <MyPrayers />
+                    <UserPrayers />
                   </StyledScrollView>
                 </SafeAreaView>
               );
